+++ api/store/model/account.go
@@ -9,11 +9,3 @@ type Account struct {
 	Password     string  `json:"password"`
 	MobileNumber *string `json:"mobileNumber"`
 }
-
-type AccountInfo struct {
-	Email       string   `json:"email"`
-	FirstName   string   `json:"firstName"`
-	LastName    string   `json:"lastName"`
-	RoleName    string   `json:"roleName"`
-	Permissions []string `json:"permissions"`
-}
