+++ api/store/model/authenticate_response.go
@@ -1,10 +1,5 @@
 package model
 
-type Credentials struct {
-	Email    string `json:"email"`
-	Password string `json:"password"`
-}
-
 type AuthenticateResponse struct {
 	AccountInfo AccountInfo `json:"accountInfo"`
 	Token       Token       `json:"token"`
