+++ api/store/model/credentials.go
@@ -1,0 +1,6 @@
+package model
+
+type Credentials struct {
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
